backend/requirements.txt
alchemical==1.0.2
alembic==1.13.2
apifairy==1.4.0
apispec==6.6.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
webargs==8.4.0
Werkzeug==3.0.3
